# docs/swagger.yaml
definitions:
  entity.Field:
    properties:
      comment:
        type: string
      id:
        type: integer
      name:
        type: string
      nullable:
        type: boolean
      table:
        $ref: '#/definitions/entity.Table'
      tableID:
        type: integer
      type:
        type: string
    type: object
  entity.Source:
    properties:
      alive:
        type: boolean
      host:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      port:
        type: integer
      schema:
        type: string
      tables:
        items:
          $ref: '#/definitions/entity.Table'
        type: array
      title:
        type: string
      type:
        description: postgresql, mysql, oracle, innodb, etc
        type: string
      username:
        type: string
    type: object
  entity.Table:
    properties:
      comment:
        type: string
      fields:
        items:
          $ref: '#/definitions/entity.Field'
        type: array
      id:
        type: integer
      name:
        type: string
      sourceID:
        type: integer
    type: object
  entity.User:
    properties:
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  routes.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  routes.SourceRequest:
    properties:
      host:
        type: string
      name:
        type: string
      password:
        type: string
      port:
        type: integer
      schema:
        type: string
      title:
        type: string
      type:
        type: string
      username:
        type: string
    required:
    - host
    - name
    - password
    - port
    - schema
    - title
    - type
    - username
    type: object
  routes.UserRequest:
    properties:
      name:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: Authorization with help username and password
      parameters:
      - description: Username and password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.LoginRequest'
      summary: Login
      tags:
      - Security
  /api/v1/logout:
    post:
      consumes:
      - application/json
      description: Logout from account
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Logout
      tags:
      - Security
  /api/v1/sources:
    get:
      consumes:
      - application/json
      description: Paginated Source List
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Source'
            type: array
      summary: List Source
      tags:
      - Sources
    post:
      consumes:
      - application/json
      description: Create entity
      parameters:
      - description: Source Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.SourceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Source'
      summary: Create Source
      tags:
      - Sources
  /api/v1/sources/:id:
    delete:
      consumes:
      - application/json
      description: Remove By ID
      parameters:
      - description: Source ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Source'
      summary: Remove Source
      tags:
      - Sources
    get:
      consumes:
      - application/json
      description: Get By ID
      parameters:
      - description: Source ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Source'
      summary: Detail Source
      tags:
      - Sources
    put:
      consumes:
      - application/json
      description: Update entity
      parameters:
      - description: Source ID
        in: path
        name: id
        required: true
        type: integer
      - description: Source Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.SourceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Source'
      summary: Update Source
      tags:
      - Sources
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: Paginated User List
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.User'
            type: array
      summary: List User
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create entity
      parameters:
      - description: User Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
      summary: Create User
      tags:
      - Users
  /api/v1/users/:id:
    delete:
      consumes:
      - application/json
      description: Remove By ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
      summary: Remove User
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Get By ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
      summary: Detail User
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update entity
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
      summary: Update User
      tags:
      - Users
swagger: "2.0"